on crud x2

+++ Java/contacts-app/src/main/java/com/contacts/app/App.java
@@ -8,15 +8,23 @@ import java.sql.SQLException;
 public class App {
     public static void main(String[] args) throws SQLException {
         final String TABLE_NAME = "personas";
-        Contact contact = new Contact("Carolina", 19, "caro@gmail.com");
-        contact.setAge(29);
+        Contact caro = new Contact("Carolina", 29, "caro@gmail.com");
+        Contact juan = new Contact("Juan Cruz", 27, "juanczmt@gmail.com");
 
         DataBase dataBase = new DataBase();
         dataBase.printAllContacts(TABLE_NAME);
-        Contact c = dataBase.findByEmail(TABLE_NAME, "juanczmt@gmail.com");
+        Contact c = dataBase.findByEmail(TABLE_NAME, juan.getEmail());
         System.out.println(c);
-        dataBase.deleteContact(TABLE_NAME, c);
+        if (c != null) {
+            dataBase.deleteContact(TABLE_NAME, c);
+        } else {
+            dataBase.saveContact(TABLE_NAME, juan);
+        }
         System.out.println("-----");
+        dataBase.printAllContacts(TABLE_NAME);
+
+        System.out.println("-----");
+        dataBase.updateEmail(TABLE_NAME, juan, "juan@gmail.com");
         dataBase.printAllContacts(TABLE_NAME);
     }
 }
